create component for tiles, fix overflow.
